derFreitag: TMP: allow str values
While migrating text fields from zope.schema.Text and TextLine to
RichText, some of them are stored not as a RichTextValue nor as a
unicode string, but rather as a plain encoded string.

This last case is what's happening on www.freitag.de and doesn't allow
for a graceful upgrade to a RichTextValue.

This commit fixes it.